# Conflicts:
#	lib/shader.c
#	lib/spritepack.c

+++ lib/renderbuffer.c
@@ -15,6 +15,7 @@ renderbuffer_initrender(struct render *r) {
 	R = r;
 }
 
+// 添加一个quad，返回1表示已经满了，无法继续添加
 int
 renderbuffer_add(struct render_buffer *rb, const struct vertex_pack vb[4], uint32_t color, uint32_t additive) {
 	if (rb->object >= MAX_COMMBINE) {
@@ -41,6 +42,7 @@ renderbuffer_add(struct render_buffer *rb, const struct vertex_pack vb[4], uint3
 
 }
 
+// 返回0表示可以更新，返回1表示不能更新
 static int
 update_tex(struct render_buffer *rb, int id) {
 	if (rb->object == 0) {
@@ -133,7 +135,7 @@ drawpolygon(struct render_buffer *rb, struct sprite_pack *pack, struct pack_poly
 		ARRAY(struct vertex_pack, vb, pn);
 
 		uv_t * texture_coord = OFFSET_TO_POINTER(uv_t, pack, p->texture_coord);
-		int16_t * screen_coord = OFFSET_TO_POINTER(int16_t, pack, p->screen_coord);
+		int32_t * screen_coord = OFFSET_TO_POINTER(int32_t, pack, p->screen_coord);
 
 		for (j=0;j<pn;j++) {
 			int xx = screen_coord[j*2+0];
